+++ lib/repositories/Interface_repos/activity/activity_repository.dart
@@ -23,7 +23,7 @@ abstract class ActivityRepository {
 
   Future<ActivityStudentSubmissionsData> getStudentSubmissions(int activityId);
 
-  Future<bool> submissionGrading(int submissionId, int grade);
+  Future<bool> submissionGrading(int submissionId, double grade);
 
   Future<bool> removeGrade(int submissionId);
