+++ lib/repositories/Interface_repos/activity/activty_datasource.dart
@@ -23,7 +23,7 @@ abstract class ActivityDataSource {
 
   Future<ActivityStudentSubmissionsData> getStudentSubmissions(int activityId);
 
-  Future<bool> submissionGrading(int submissionId, int grade);
+  Future<bool> submissionGrading(int submissionId, double grade);
 
   Future<bool> removeGrade(int submissionId);
 }
